// client/src/utils/notes-statuses.ts
export const enum NotesStatuses {
    CREATED = "Создана",
    REVISING = "На доработке",
    APPROVING = "На согласовании",
    SIGNING = "На подписании",
    REGISTRATING = "На регистрации",
    REGISTERED = 'Зарегистрирована',
    APPROVED = "Согласована",
    SIGNED = "Подписана",
}